All - Fix LC_MONETARY with '@' For All
