+++ NickvisionMoney.GNOME/Views/NewAccountDialog.cs
@@ -1,6 +1,7 @@
 using Nickvision.Aura.Keyring;
 using NickvisionMoney.GNOME.Helpers;
 using NickvisionMoney.Shared.Controllers;
+using NickvisionMoney.Shared.Helpers;
 using NickvisionMoney.Shared.Models;
 using System;
 using System.IO;
@@ -217,7 +218,7 @@ public partial class NewAccountDialog : Adw.Window
         _accountTypeRow.SetSelected(0);
         _incomeButton.SetActive(true);
         _transactionRemindersRow.SetSelected(0);
-        _reportedCurrencyLabel.SetLabel($"{_("Your system reported that your currency is")}\n<b>{_controller.ReportedCurrencyString}</b>");
+        _reportedCurrencyLabel.SetLabel($"{_("Your system reported that your currency is")}\n<b>{CultureHelpers.ReportedCurrencyString}</b>");
     }
     
     /// <summary>
