+++ NickvisionMoney.Shared/Controllers/AccountSettingsDialogController.cs
@@ -63,32 +63,6 @@ public class AccountSettingsDialogController
     }
 
     /// <summary>
-    /// The system reported currency string (Ex: "$ (USD)")
-    /// </summary>
-    public string ReportedCurrencyString
-    {
-        get
-        {
-            var lcMonetary = Environment.GetEnvironmentVariable("LC_MONETARY");
-            if (lcMonetary != null && lcMonetary.Contains(".UTF-8"))
-            {
-                lcMonetary = lcMonetary.Remove(lcMonetary.IndexOf(".UTF-8"), 6);
-            }
-            else if (lcMonetary != null && lcMonetary.Contains(".utf8"))
-            {
-                lcMonetary = lcMonetary.Remove(lcMonetary.IndexOf(".utf8"), 5);
-            }
-            if (lcMonetary != null && lcMonetary.Contains('_'))
-            {
-                lcMonetary = lcMonetary.Replace('_', '-');
-            }
-            var culture = new CultureInfo(!string.IsNullOrEmpty(lcMonetary) ? lcMonetary : CultureInfo.CurrentCulture.Name, true);
-            var region = new RegionInfo(!string.IsNullOrEmpty(lcMonetary) ? lcMonetary : CultureInfo.CurrentCulture.Name);
-            return $"{culture.NumberFormat.CurrencySymbol} ({region.ISOCurrencySymbol})";
-        }
-    }
-
-    /// <summary>
     /// Updates the Metadata object
     /// </summary>
     /// <param name="name">The new name of the account</param>
